extracted router setup
