+++ cmd/shortener/main.go
@@ -1,14 +1,10 @@
 package main
 
 import (
-	"github.com/gin-gonic/gin"
-	"github.com/khodya/go-urlshortener/internal/app"
+	"github.com/khodya/go-urlshortener/internal/router"
 )
 
 func main() {
-	router := gin.Default()
-	router.GET("/:id", app.Unfold)
-	router.POST("/", app.Fold)
-
+	router := router.SetupRouter()
 	router.Run("localhost:8080")
 }
